abstracted commands with /utils/ directory
