+++ bot/package.json
@@ -5,12 +5,14 @@
   "main": "index.js",
   "scripts": {
     "start": "node .",
+    "dev": "nodemon .",
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "author": "Carter Costic",
   "license": "ISC",
   "devDependencies": {
-    "eslint": "^8.6.0"
+    "eslint": "^8.6.0",
+    "nodemon": "^2.0.15"
   },
   "dependencies": {
     "@discordjs/builders": "^0.11.0",
